Update fishing icon and add new aquatic life icon
