+++ app/client/src/components/pages/AquaticLife/index.js
@@ -7,9 +7,10 @@ import type { RouteProps } from 'routes.js';
 import Page from 'components/shared/Page';
 import NavBar from 'components/shared/NavBar';
 import LocationSearch from 'components/shared/LocationSearch';
-import DrinkingWaterIcon from 'components/shared/Icons/DrinkingWaterIcon';
 import SwimmingIcon from 'components/shared/Icons/SwimmingIcon';
-import FishingIcon from 'components/shared/Icons/FishingIcon';
+import EatingFishIcon from 'components/shared/Icons/EatingFishIcon';
+import AquaticLifeIcon from 'components/shared/Icons/AquaticLifeIcon';
+import DrinkingWaterIcon from 'components/shared/Icons/DrinkingWaterIcon';
 import DisclaimerModal from 'components/shared/DisclaimerModal';
 import ShowLessMore from 'components/shared/ShowLessMore';
 // styled components
@@ -44,7 +45,7 @@ function AquaticLife({ ...props }: Props) {
       <NavBar title="Explore Topics" />
 
       <StyledTopic>
-        <FishingIcon />
+        <AquaticLifeIcon />
         <p>Aquatic Life</p>
       </StyledTopic>
 
@@ -91,7 +92,7 @@ function AquaticLife({ ...props }: Props) {
             Swimming
           </TopicButtonLink>
           <TopicButtonLink to="/eating-fish">
-            <FishingIcon />
+            <EatingFishIcon />
             Eating Fish
           </TopicButtonLink>
           <TopicButtonLink to="/drinking-water">
