+++ app/client/src/components/pages/DrinkingWater/index.js
@@ -7,9 +7,10 @@ import type { RouteProps } from 'routes.js';
 import Page from 'components/shared/Page';
 import NavBar from 'components/shared/NavBar';
 import LocationSearch from 'components/shared/LocationSearch';
-import DrinkingWaterIcon from 'components/shared/Icons/DrinkingWaterIcon';
 import SwimmingIcon from 'components/shared/Icons/SwimmingIcon';
-import FishingIcon from 'components/shared/Icons/FishingIcon';
+import EatingFishIcon from 'components/shared/Icons/EatingFishIcon';
+import AquaticLifeIcon from 'components/shared/Icons/AquaticLifeIcon';
+import DrinkingWaterIcon from 'components/shared/Icons/DrinkingWaterIcon';
 // styled components
 import { StyledText, StyledTopic } from 'components/shared/Topics';
 import {
@@ -71,11 +72,11 @@ function DrinkingWater({ ...props }: Props) {
             Swimming
           </TopicButtonLink>
           <TopicButtonLink to="/aquatic-life">
-            <FishingIcon />
+            <AquaticLifeIcon />
             Aquatic Life
           </TopicButtonLink>
           <TopicButtonLink to="/eating-fish">
-            <FishingIcon />
+            <EatingFishIcon />
             Eating Fish
           </TopicButtonLink>
         </StyledButtons>
