+++ app/client/src/components/pages/EatingFish/index.js
@@ -7,9 +7,10 @@ import type { RouteProps } from 'routes.js';
 import Page from 'components/shared/Page';
 import NavBar from 'components/shared/NavBar';
 import LocationSearch from 'components/shared/LocationSearch';
-import DrinkingWaterIcon from 'components/shared/Icons/DrinkingWaterIcon';
 import SwimmingIcon from 'components/shared/Icons/SwimmingIcon';
-import FishingIcon from 'components/shared/Icons/FishingIcon';
+import EatingFishIcon from 'components/shared/Icons/EatingFishIcon';
+import AquaticLifeIcon from 'components/shared/Icons/AquaticLifeIcon';
+import DrinkingWaterIcon from 'components/shared/Icons/DrinkingWaterIcon';
 import DisclaimerModal from 'components/shared/DisclaimerModal';
 import ShowLessMore from 'components/shared/ShowLessMore';
 // styled components
@@ -44,7 +45,7 @@ function EatingFish({ ...props }: Props) {
       <NavBar title="Explore Topics" />
 
       <StyledTopic>
-        <FishingIcon />
+        <EatingFishIcon />
         <p>Eating Fish</p>
       </StyledTopic>
 
@@ -90,7 +91,7 @@ function EatingFish({ ...props }: Props) {
             Swimming
           </TopicButtonLink>
           <TopicButtonLink to="/aquatic-life">
-            <FishingIcon />
+            <AquaticLifeIcon />
             Aquatic Life
           </TopicButtonLink>
           <TopicButtonLink to="/drinking-water">
