+++ app/client/src/components/shared/Icons/AquaticLifeIcon.js
@@ -1,0 +1,32 @@
+// @flow
+
+import React from 'react';
+
+type Props = {
+  height: string,
+};
+
+function AquaticLifeIcon({ height = '5rem' }: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      x="0"
+      y="0"
+      viewBox="0 0 54 54"
+      style={{ height }}
+    >
+      <path
+        fill="#fff"
+        d="M9.19,17.84c2.89,2.27,6,4.79,6.88,8.35.85,3.22-.24,6.75.78,9.92a11.61,11.61,0,0,0,3.7,5.07c1.6,1.38,3.39,2.53,5,3.89-2.65-3.07-5.34-6.24-6.73-10s-1.24-8.45,1.42-11.5a26.78,26.78,0,0,1-5.38-3.76,19.12,19.12,0,0,1,7.36-4.54A21.43,21.43,0,0,0,17.83,8a18.12,18.12,0,0,1,4.4,5.11,3.55,3.55,0,0,0,1.34,1.59.49.49,0,0,0,.6,0,.56.56,0,0,0,.06-.36,41.83,41.83,0,0,0-.45-4.17l.85,4.35a15.59,15.59,0,0,1,6,.27,11.64,11.64,0,0,1,1.49-4.84l-.19,4.27A1.46,1.46,0,0,0,32,15a1.47,1.47,0,0,0,.72.55L35.91,17a24.48,24.48,0,0,0,.43-5.6c.29-.16.51.3.52.63L37,17.76a46.27,46.27,0,0,0,6.87-2.23,16.33,16.33,0,0,0,.09,8c-1.71.24-3.23-1-4.91-1.45a5.25,5.25,0,0,0-4,.63c-.47.28-1.21.73-1.22,1.32a20.5,20.5,0,0,0,.39,2.8,9.18,9.18,0,0,1-.19,4.83c-1.1,2.6-4.29,3.66-5.84,6-1.21,1.84-.89,3.1-.63,5.28,2.14-2.52,2.74-5,5.57-6.73A10.28,10.28,0,0,1,39,35.15a31.76,31.76,0,0,0,5.93,0,4,4,0,0,1-3,3.41,12.77,12.77,0,0,1-4.82.24,10.41,10.41,0,0,0-4.78.44c-2.37,1-3.7,3.49-4.78,5.82-.19.4-.44.86-.87.92a1.33,1.33,0,0,1-.65-.13c-4.12-1.61-8.45-3.7-10.44-7.64a13.48,13.48,0,0,1-1.18-7c.12-1.82.48-3.68.05-5.46C13.68,22.67,10.79,20.6,9.19,17.84ZM23.1,34.23A32.3,32.3,0,0,0,25.88,43c.07.17.24.38.4.29a35.38,35.38,0,0,1-1.54-7.48,11.61,11.61,0,0,1,.44-5.31,12.22,12.22,0,0,1,4.54-5.14.94.94,0,0,0,.47-.59,14.54,14.54,0,0,1-5.84,0,.28.28,0,0,0-.28,0A.25.25,0,0,0,24,25c-.19,1-.38,2.09-.55,3.13A24.2,24.2,0,0,0,23,33C23,33.39,23.05,33.81,23.1,34.23Zm9.29-6.44A9,9,0,0,0,32.18,26a.33.33,0,0,0-.11-.21c-.21-.17-2.92,3-3.25,3.31a9.64,9.64,0,0,0-2,2.68A8.7,8.7,0,0,0,26.29,35c-.05,1.66,0,3.33.05,5a14.85,14.85,0,0,0,.37,2.88c.2,0,.22-.86.21-1.06-.05-1.19-.13-1-.18-2.15A9.4,9.4,0,0,1,27,36.27a10.13,10.13,0,0,1,2.77-3.69,8,8,0,0,0,2.53-3.82A5.23,5.23,0,0,0,32.39,27.79Zm-11.9,4.29s-.08,0-.11.07a.46.46,0,0,0,0,.26A22.59,22.59,0,0,0,23.6,40.9c.61,1.09,1.25,2.17,1.9,3.24.07.12.23.26.34.16a19.63,19.63,0,0,0-1.42-3c-.51-1.06-1-2.1-1.51-3.16S22,36,21.54,34.9c-.21-.54-.4-1.09-.58-1.64C20.92,33.13,20.69,32.13,20.49,32.08Zm.76-13.45a.85.85,0,0,0-.12-.47c-.42-.67-1.72-.56-1.83.3a1,1,0,0,0,.49.93,1,1,0,0,0,1.16-.1A1,1,0,0,0,21.25,18.63Z"
+      />
+    </svg>
+  );
+}
+
+// 'height' defaultProp is set here just to satisfy flow,
+// even though its passed as a default function param
+AquaticLifeIcon.defaultProps = {
+  height: '5rem',
+};
+
+export default AquaticLifeIcon;
