+++ app/client/src/components/shared/Icons/EatingFishIcon.js
@@ -1,0 +1,44 @@
+// @flow
+
+import React from 'react';
+
+type Props = {
+  height: string,
+};
+
+function EatingFishIcon({ height = '5rem' }: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      x="0"
+      y="0"
+      viewBox="0 0 54 54"
+      style={{ height }}
+    >
+      <path
+        fill="#fff"
+        d="M49.54,26.75V12.06c0-.35.12-.82-.3-1s-.59.34-.8.61a10.37,10.37,0,0,0-1.61,3.43,51,51,0,0,0-1.44,12.24,1.61,1.61,0,0,0,.81,1.53,2.18,2.18,0,0,1,1,2.32c-.71,3.2-.25,6.43-.36,9.64a2.88,2.88,0,0,0,.24,1.48,1.19,1.19,0,0,0,1.2.69,1,1,0,0,0,1-.7,5.84,5.84,0,0,0,.16-1.44Q49.54,33.78,49.54,26.75Z"
+      />
+      <path
+        fill="#fff"
+        d="M23.56,33a1.41,1.41,0,0,0-.72-.36,11.27,11.27,0,0,1-4.5-2.13A11,11,0,0,1,15.82,28a1.14,1.14,0,0,1,0-1.44,10.33,10.33,0,0,1,1.62-1.81,11.66,11.66,0,0,1,5.62-2.91.5.5,0,0,0,.33-.2,18.43,18.43,0,0,1,4-2.7c.24-.14.5-.25.78-.39V22a.42.42,0,0,0,.29.48A11.45,11.45,0,0,1,32,24.88c.4.38.77.81,1.15,1.21a19.11,19.11,0,0,1,4.87-3.23,3,3,0,0,1-.06.77,13,13,0,0,1-1.09,3.44.47.47,0,0,0,0,.46,11,11,0,0,1,1,2.95c.07.35.08.71.12,1.13a19.16,19.16,0,0,1-4.89-3.26l-.13.13A11.76,11.76,0,0,1,28.53,32a.52.52,0,0,0-.38.6V36A19.4,19.4,0,0,1,23.56,33Zm-3-4.55a1.24,1.24,0,0,0,1.24-1.24A1.27,1.27,0,0,0,20.52,26a1.25,1.25,0,1,0,0,2.49Z"
+      />
+      <path
+        fill="#fff"
+        d="M43.13,27.13A15.66,15.66,0,1,0,27.37,42.69h0A15.64,15.64,0,0,0,43.13,27.16ZM27.42,39.41a12.48,12.48,0,0,1-.55-24.95c.22,0,.44,0,.66,0a12.48,12.48,0,0,1,.56,24.95C27.87,39.42,27.64,39.41,27.42,39.41Z"
+      />
+      <path
+        fill="#fff"
+        d="M8.69,11c-.38.36-.27.83-.27,1.25v6c0,.35.16,1-.45.9s-.34-.58-.34-.9V12c0-.37.14-1-.41-1s-.4.61-.41,1v6.24c0,.38.11,1-.45,1-.71,0-.44-.63-.44-1,0-2.16.07-4.32.09-6.48,0-.26.06-.66-.35-.65s-.31.34-.33.58c-.35,2.25-.62,4.51-.87,6.71a3.35,3.35,0,0,0,1,2.77A5.55,5.55,0,0,1,7,26.79a6.31,6.31,0,0,0-.26,1.44V41.56C6.69,42.4,7.07,43,8,43s1.29-.6,1.32-1.44c0-1.27.05-2.53,0-3.79C9.23,34,9.66,30.16,9,26.39a5.9,5.9,0,0,1,1.47-5.32,3.93,3.93,0,0,0,1-3.39c-.25-1.86-.44-3.72-.68-5.59,0-.4.07-1.08-.54-1s-.33.67-.32,1c0,2.07.13,4.16.19,6.23,0,.39.07.9-.5.95s-.51-.54-.52-.9V12.27C9.06,11.85,9.26,11.35,8.69,11Z"
+      />
+    </svg>
+  );
+}
+
+// 'height' defaultProp is set here just to satisfy flow,
+// even though its passed as a default function param
+EatingFishIcon.defaultProps = {
+  height: '5rem',
+};
+
+export default EatingFishIcon;
