--- app/client/src/components/shared/Icons/FishingIcon.js
@@ -1,32 +1,0 @@
-// @flow
-
-import React from 'react';
-
-type Props = {
-  height: string,
-};
-
-function FishingIcon({ height = '5rem' }: Props) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      x="0"
-      y="0"
-      viewBox="0 0 54 54"
-      style={{ height }}
-    >
-      <path
-        fill="#fff"
-        d="M16.26,16.04c0.67-1.28,1.71-1.62,3.04-1.59c4.43,0.1,8.88-0.06,13.29,0.26c1.93,0.14,3.89,0.98,5.64,1.87c1.19,0.6,2.03,1.89,3.1,2.78c0.75,0.63,1.55,1.34,2.45,1.65c2.39,0.81,4.27,2.31,5.9,4.12c1.01,1.12,1.8,2.45,2.64,3.71c0.78,1.17-0.13,3.01-1.55,3.48c-1.36,0.45-2.66,1.1-3.98,1.67c-0.07,0.03-0.13,0.07-0.2,0.08c-1.95,0.34-3.89,0.69-5.84,1c-0.55,0.09-1.11,0.11-1.67,0.1c-0.83,0-1.21,0.26-1.56,1.16c-0.5,1.27-1.48,2.36-2.26,3.52c-0.48,0.71-0.93,1.45-1.46,2.12c-0.69,0.86-2.17,0.76-3.01,0.04c-1.82-1.58-1.82-3.63-1.58-5.71c0.12-1.01-0.21-1.25-1.12-1.24c-2.6,0.04-5.21,0.05-7.82,0.02c-1.27-0.01-2.22-1.05-2.23-2.33c-0.01-1.26-0.77-2.12-2.01-2.34c-2.58-0.45-4.57,0.56-6.32,2.37c-1.3,1.34-2.67,2.6-4.06,3.84c-0.6,0.53-1.53,0.07-1.58-0.82c-0.09-1.85-0.06-3.71-0.1-5.57c-0.03-1.31,0.21-2.5-0.69-3.87c-1.14-1.72-1.56-3.96-1.11-6.16c0.2-0.97,1.68-1.48,2.6-0.9c0.83,0.53,1.66,1.06,2.49,1.61c0.89,0.59,1.71,1.32,2.66,1.78c1.62,0.78,3.28,1.51,5.17,1.3c1.51-0.16,2.72-0.92,3.93-1.74c0.26-0.18,0.61-0.23,0.93-0.32c0.87-0.23,1.05-0.89,0.43-1.54C19,18.94,17.63,17.48,16.26,16.04zM45.06,27.16c-0.51-0.46-0.97-1.17-1.39-1.15c-0.43,0.02-0.84,0.75-1.25,1.17c0.42,0.42,0.83,1.14,1.27,1.16C44.11,28.35,44.55,27.63,45.06,27.16z"
-      />
-    </svg>
-  );
-}
-
-// 'height' defaultProp is set here just to satisfy flow,
-// even though its passed as a default function param
-FishingIcon.defaultProps = {
-  height: '5rem',
-};
-
-export default FishingIcon;
